- Padronizado Ids com o tipo long
- Alteração na versão e descrição do pacote

+++ rmdev.ibge.localidades/Contracts/IIBGELocalidades_Distrito.cs
@@ -7,15 +7,13 @@ namespace rmdev.ibge.localidades
 {
     public partial interface IIBGELocalidades
     {
-        //TODO: Ver Distritos
-
         /// <summary>
         /// Obtém o conjunto de distritos do Brasil a partir dos respectivos identificadores
         /// </summary>
         /// <param name="codigoDistritos">Um ou mais identificadores de distrito</param>
         /// <returns>Lista de distritos</returns>
         [Get("/api/v1/localidades/distritos/{codigoDistritos}")]
-        Task<List<Distrito>> BuscarDistritosAsync(params int[] codigoDistritos);
+        Task<List<Distrito>> BuscarDistritosAsync(params long[] codigoDistritos);
 
         /// <summary>
         /// Obtém o conjunto de distritos do Brasil a partir dos identificadores das Unidades da Federação
@@ -23,7 +21,7 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoUFs">Um ou mais identificadores de Unidades da Federação</param>
         /// <returns>Lista de distritos</returns>
         [Get("/api/v1/localidades/estados/{codigoUFs}/distritos")]
-        Task<List<Distrito>> BuscarDistritosPorUFAsync(params int[] codigoUFs);
+        Task<List<Distrito>> BuscarDistritosPorUFAsync(params long[] codigoUFs);
 
         /// <summary>
         /// Obtém o conjunto de distritos do Brasil a partir dos identificadores das mesorregiões
@@ -31,7 +29,7 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoMesorregioes">Um ou mais identificadores de mesorregiões</param>
         /// <returns>Lista de distritos</returns>
         [Get("/api/v1/localidades/mesorregioes/{codigoMesorregioes}/distritos")]
-        Task<List<Distrito>> BuscarDistritosPorMesorregiaoAsync(params int[] codigoMesorregioes);
+        Task<List<Distrito>> BuscarDistritosPorMesorregiaoAsync(params long[] codigoMesorregioes);
 
         /// <summary>
         /// Obtém o conjunto de distritos do Brasil a partir dos identificadores das microrregiões
@@ -39,7 +37,7 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoMicrorregioes">Um ou mais identificadores de microrregiões</param>
         /// <returns>Lista de distritos</returns>
         [Get("/api/v1/localidades/microrregioes/{codigoMicrorregioes}/distritos")]
-        Task<List<Distrito>> BuscarDistritosPorMicrorregiaoAsync(params int[] codigoMicrorregioes);
+        Task<List<Distrito>> BuscarDistritosPorMicrorregiaoAsync(params long[] codigoMicrorregioes);
 
         /// <summary>
         /// Obtém o conjunto de distritos do Brasil a partir dos identificadores dos municípios
@@ -47,7 +45,7 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoMunicipios">Um ou mais identificadores de municipios</param>
         /// <returns>Lista de distritos</returns>
         [Get("/api/v1/localidades/municipios/{codigoMunicipios}/distritos")]
-        Task<List<Distrito>> BuscarDistritosPorMunicipioAsync(params int[] codigoMunicipios);
+        Task<List<Distrito>> BuscarDistritosPorMunicipioAsync(params long[] codigoMunicipios);
 
         /// <summary>
         /// Obtém o conjunto de distritos do Brasil a partir dos identificadores das regiões imediatas
@@ -55,7 +53,7 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoRegiaoImediata">Um ou mais identificadores de regiões imediatas</param>
         /// <returns>Lista de distritos</returns>
         [Get("/api/v1/localidades/regioes-imediatas/{codigoRegiaoImediata}/distritos")]
-        Task<List<Distrito>> BuscarDistritosPorRegiaoImediataAsync(params int[] codigoRegiaoImediata);
+        Task<List<Distrito>> BuscarDistritosPorRegiaoImediataAsync(params long[] codigoRegiaoImediata);
 
         /// <summary>
         /// Obtém o conjunto de distritos do Brasil a partir dos identificadores das regiões intermediárias
@@ -63,7 +61,7 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoRegiaoIntermediaria">Um ou mais identificadores de regiões intermediárias</param>
         /// <returns>Lista de distritos</returns>
         [Get("/api/v1/localidades/regioes-intermediarias/{codigoRegiaoIntermediaria}/distritos")]
-        Task<List<Distrito>> BuscarDistritosPorRegiaoIntermediariaAsync(params int[] codigoRegiaoIntermediaria);
+        Task<List<Distrito>> BuscarDistritosPorRegiaoIntermediariaAsync(params long[] codigoRegiaoIntermediaria);
 
         /// <summary>
         /// Obtém o conjunto de distritos do Brasil a partir dos identificadores das regiões
@@ -71,26 +69,26 @@ namespace rmdev.ibge.localidades
         /// <param name="codigoMacrorregiao">Um ou mais identificadores de regiões</param>
         /// <returns>Lista de distritos</returns>
         [Get("/api/v1/localidades/regioes/{codigoMacrorregiao}/distritos")]
-        Task<List<Distrito>> BuscarDistritosPorMacrorregiaoAsync(params int[] codigoMacrorregiao);
+        Task<List<Distrito>> BuscarDistritosPorMacrorregiaoAsync(params long[] codigoMacrorregiao);
 
         /// <summary>
-        /// Obtém um único distritos do Brasil a partir do identificador
+        /// Obtém um único distrito do Brasil a partir do identificador
         /// </summary>
         /// <param name="codigoDistrito">Identificador de distrito</param>
         /// <returns>Dados do distrito</returns>
-        public async Task<Distrito?> BuscarDistritoAsync(int codigoDistrito)
+        public async Task<Distrito?> BuscarDistritoAsync(long codigoDistrito)
         {
             var distritos = await BuscarDistritosAsync(codigoDistrito);
             return distritos.FirstOrDefault();
         }
 
         
-        //TODO: Ver Mesorregiões 
-        //TODO: Ver Microrregiões 
-        //TODO: Ver Regiões imediatas
-        //TODO: Ver Regiões integradas de desenvolvimento
-        //TODO: Ver Regiões intermediárias
-        //TODO: Ver Regiões metropolitanas
-        //TODO: Ver Subdistritos
+        //TODO: Ler documentação e implementar consultas por mesorregiões 
+        //TODO: Ler documentação e implementar consultas por microrregiões 
+        //TODO: Ler documentação e implementar consultas por regiões imediatas
+        //TODO: Ler documentação e implementar consultas por regiões integradas de desenvolvimento
+        //TODO: Ler documentação e implementar consultas por regiões intermediárias
+        //TODO: Ler documentação e implementar consultas por regiões metropolitanas
+        //TODO: Ler documentação e implementar consultas por subdistritos
     }
 }
